[REFACTOR]:added dynamic roles to the protected routes
