+++ src/routes/AdminProtectedRoute.jsx
@@ -1,6 +1,6 @@
 import { useSelector } from 'react-redux';
 import { Navigate, Outlet } from 'react-router-dom';
-import { roles } from '../constants/roles';
+import { roles } from '../constants/constants';
 
 const AdminProtectedRoute = () => {
   const { user } = useSelector((state) => state.auth);
